% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeBF_function.R
\name{visualizeBF}
\alias{visualizeBF}
\title{Function for exploring Bayes Factors.}
\usage{
visualizeBF(data, pointsize = 0.001, scale = 0.707, plot = 1)
}
\arguments{
\item{data}{A vector of data}

\item{pointsize}{Width of points for numerical integration, defaults to .001.}

\item{scale}{Scale parameter for Cauchy prior, defaults to .707}

\item{plot}{The type of plot to produce. 1=BF as weighted likelihood, 2=BF as comparison of likelihood for mu=0 under prior and posterior distributions.}
}
\value{
Returns a list with the following components:  
\itemize{
  \item \code{BF10}    The Bayes Factor for alternative hypothesis relative to null.  \cr 
  \item \code{BF01}    The Bayes Factor for the null hypothesis relative to the alternative \cr 
 }
}
\description{
This function plots the prior, likelihood, and posterior distribution of the sample mean. It reports the MAP (maximum a posteriori) and EAP (expected a posteriori) estimates of the mean as well as the 95\% credible interval around the MAP.
}
\examples{
set.seed(1)
data <- rnorm(n=50, mean=.3, sd=1)
visualizeBF(data, plot=1)
visualizeBF(data, plot=2)
}
