% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeBF_function.R
\name{visualizeBF}
\alias{visualizeBF}
\title{Function for exploring Bayes Factors.}
\usage{
visualizeBF(data, pointsize = 0.001, scale = 0.707, plot = 1)
}
\arguments{
\item{data}{A vector of data}

\item{pointsize}{Width of points for numerical integration, defaults to .001.}

\item{scale}{Scale parameter for Cauchy prior, defaults to .707}

\item{plot}{The type of plot to produce. 1=BF as weighted likelihood, 2=BF as comparison of densities for mu=0 under prior and posterior distributions.}
}
\value{
Returns a list with the following components:  
\itemize{
  \item \code{L.m0}    The likelihood evaluated at mu=0; the L of the null model (m0) \cr
  \item \code{L.m1}    The likelihood integrated over the whole space weighted by m1 prior; L of the alternative model (m1). \cr
  \item \code{prior.mu0} The density of the prior distribution evaluated at mu=0 (m1). \cr
  \item \code{posterior.m0} The density the posterior distribution evaluated at mu=0 (m0). \cr
  \item \code{BF10}    The Bayes Factor for alternative hypothesis relative to null.  \cr 
  \item \code{BF01}    The Bayes Factor for the null hypothesis relative to the alternative \cr 
  \item \code{figure}  Object containing the plot \cr
 }
}
\description{
Function for exploring Bayes Factors.
}
\examples{
set.seed(1)
data <- rnorm(n=50, mean=.3, sd=1)
visualizeBF(data, plot=1)
visualizeBF(data, plot=2)

a <- visualizeBF(data, plot=1)

# extract density of the prior distribution at mu=0
a$prior.mu0

# calculate BF10 as likelihood ratio
a$L.m1 / a$L.m0

# alternative method for calculating BF10 as density ratio
a$prior.mu0 / a$posterior.mu0

# view the plot
a$figure
}
